+++ custom/justech_l10n_do_ncf/__init__.py
@@ -1,3 +1,7 @@
 from . import models
 from . import services
 from . import wizards
+
+
+def post_init_hook(env):
+    env["justech.do.purchase.emission.config"].ensure_configs_for_companies()